refactor: use providers directory with index for better organization
